+++ h_project/voice_processing/voice_preprocessor.py
@@ -1,0 +1,30 @@
+import numpy
+import pandas as pd
+import os
+
+
+class VoicePreprocessor():
+
+    """
+    This function
+    """
+
+    def __init__(self):
+
+        pass
+
+    def process_folder(self, folder):
+        """
+
+        :param folder:
+        :return:
+        """
+
+        pass
+
+    def segment_raw_audio(self, raw_audio, fs):
+
+        
+
+
+        pass
